feat: added deep copy on json parser
